requirements.txt
fastapi
uvicorn
litellm
python-dotenv
httpx
rich
pandas
pymongo>=4.10.0